Switch to opencv2 everywhere

+++ atruck/GreenFlag.py
@@ -73,7 +73,8 @@ class GreenFlag(object):
 
         self._im_labels, n_labels = measure.label(mask, background=0,
                                                   return_num=True)
-        self._labels_sizes = np.bincount(self._im_labels.ravel())
+        self._labels_sizes = np.bincount(
+            self._im_labels.astype(np.uint32).ravel())
         im_size = mask.shape[0] * mask.shape[1]
         for label in xrange(1, n_labels + 1):
             if not self._valid_size(label, im_size):
